+++ test/e2e/spec/account/medcryptor.feature
@@ -1,6 +1,5 @@
 Feature: MedCryptor users
 
-    @wip
     Scenario: MedCryptor users have extra data
         Given I create a MedCryptor account
         Then  I can edit specialization, medical ID, country and role
